Added generic one-to-many / many-to-many fetching
getForeignKeyFieldNameOn allows you to fetch the FK fieldName
of a model A that has a FK-field to a model B.
If throws an exception if there is no such field.

This is the basis for the implementation of "getManyFromOne".

Additionally there's "unpackFromJoinModel".
This macro allows unpacking a fetched many-to-many relationship.
This macro is the basis for the "getManyToMany" function. Used elsewhere