#23 Second implementation of connection pooling
This time you have a fixed size base pool and a growing BURSTPOOL.
The BURSTPOOL gets initialized when the basepool runs out of connections.
The Burstpool then has a lifetime of 30 minutes, which
gets extended by 5 seconds every time a new connection
is borrowed from it.

This also provides a template for making a query with.