FIX breakage in timestampModel
Constructor has a cache table where it temporarily stores "cleaned up" versions of the Models fed into it. Likely after scanning for the "defaults" pragma and adding everything into the table accordingly.

Somehow, the second time round when it eats the type via "implDefaults" it fails to properly do the conversion due to some change in the nim language.
The flag defined here disables that change temporarily.

Signed-off-by: Philipp Doerner <philippmdoerner@web.de>