REFACTOR password hashing to use nimword
Previously password hashing was done using a custom reimplementation
of pbkdf2_hmac_sha256.
Now it is using nimword instead.
Nimword uses a slightly different stringformat though.
It expects the salts to be base64 encoded and a $ at the beginning.
It also expects removal of trailing "=" characters.
Thus fix_user_hashes was added to update the database accordingly.

The dockerfiles were updated to download
libsodium as the new default hashing algorithm
has been changed to be argon2id, which requires libsodium